Move to a new Makefile that's installed through npm (#346)
Now using https://github.com/rowanmanning/make, which is basically the
same as before except it's managed through npm. This reduces boilerplate
code in the repo.

To get the tests running again I had to append each file with .test.js,
this also allows us to clearly separate tests from helpers and mocks.
