+++ package.json
@@ -40,6 +40,7 @@
     "semver": "^5.4.1"
   },
   "devDependencies": {
+    "@rowanmanning/make": "^1.0.1",
     "eslint": "^3.18.0",
     "mocha": "^3.2.0",
     "mockery": "^2.0.0",
