+++ test/integration/cli/actions-click-element.test.js
