+++ test/integration/cli/actions-screen-capture.test.js
