+++ test/integration/cli/add-rule.test.js
