+++ test/integration/cli/config.test.js
