+++ test/integration/cli/environment.test.js
