+++ test/integration/cli/hide-elements.test.js
