+++ test/integration/cli/ignore.test.js
