+++ test/integration/cli/output.test.js
