+++ test/integration/cli/reporter-json.test.js
