+++ test/integration/cli/root-element.test.js
