+++ test/integration/cli/screen-capture.test.js
