+++ test/integration/cli/standard.test.js
