+++ test/integration/cli/timeout.test.js
