+++ test/integration/cli/wait.test.js
