+++ test/unit/lib/reporter.test.js
