+++ test/unit/mock/window.mock.js
